Move functions to non-index file to make them testable
